requirements.txt
# Health Data Export System Requirements

# Core dependencies
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
pandas==2.1.3
pymysql==1.1.0
python-dotenv==1.0.0
schedule==1.2.0

# Optional: For advanced features
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
httpx==0.25.2

# Development dependencies
pytest==7.4.3
black==23.11.0
flake8==6.1.0
